requirements.txt
strands-agents
strands-agents-tools
boto3
diagrams